Refactoring
Printing the logo directly instead of saving it to a variable first.
Changed some comments to be shorter and more readable. 
Corrected position_dict to be PEP compliant.
Changed empty_board_generator at line 77 to use _ instead of x and y.
Changed line 128 from "if player1_turn == False:" to  "if not player1_turn:" since it's a boolean and the if statement simply wants that.